Add some improvement in Services

+++ EndPoint/Controllers/UserController.cs
@@ -1,9 +1,11 @@
-﻿using Application.Command.DTO.User;
+﻿using System.Threading.Tasks;
+using Application.Command.DTO.User;
 using Application.Command.Services.User;
 using Application.Command.Utilities;
 using Domain.Entity;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Identity.Client;
 
 namespace EndPoint.Controllers
 {
@@ -13,14 +15,18 @@ namespace EndPoint.Controllers
     {
         private readonly IUserService _userService;
         private readonly IUserServiceQuery _userServiceQuery;
-        public UserController(IUserService userService , IUserServiceQuery userServiceQuery)
+        public UserController(IUserService userService, IUserServiceQuery userServiceQuery)
         {
             _userService = userService;
             _userServiceQuery = userServiceQuery;
         }
-        public IActionResult Register(string name, string username, string email, string password)
+        public async Task<IActionResult> Register(string name, string username, string email, string password)
         {
-            var result = _userService.Register(new Application.Command.DTO.User.SignUpDTO
+            if (string.IsNullOrWhiteSpace(name) || string.IsNullOrWhiteSpace(username) || string.IsNullOrWhiteSpace(email) || string.IsNullOrWhiteSpace(password))
+            {
+                return BadRequest("Informations can not be empty!");
+            }
+            var result = await _userService.Register(new Application.Command.DTO.User.SignUpDTO
             {
                 Fullname = name,
                 Password = password,
@@ -29,17 +35,34 @@ namespace EndPoint.Controllers
             }
             );
 
-            return Ok(result);
+            if (result.Status == Status.Success)
+            {
+                return Ok(result);
+            }
+
+            return BadRequest(result);
         }
         [HttpPost("RemoveAccount")]
-        public IActionResult RemoveAccount(string username, string password)
+        public async Task<IActionResult> RemoveAccount(string username, string password)
         {
-            var result = _userService.DeleteAccount(new Application.Command.DTO.User.RemoveAccountDTO
+            if (string.IsNullOrWhiteSpace(username) || string.IsNullOrWhiteSpace(password))
+            {
+                return BadRequest("Your username or password can not be empty!");
+            }
+            var result = await _userService.DeleteAccount(new Application.Command.DTO.User.RemoveAccountDTO
             {
                 Username = username,
                 Password = password
             });
-            return Ok(result);
+            if (result.Status == Status.Success)
+            {
+                return Ok(result);
+            }
+            else
+            {
+                return BadRequest(result);
+            }
+
         }
 
         [HttpPost("GetAll")]
@@ -49,9 +72,9 @@ namespace EndPoint.Controllers
         }
 
         [HttpPost("Update")]
-        public IActionResult Update(string fullname, string username, string password, string email)
+        public async Task<IActionResult> Update(string fullname, string username, string password, string email)
         {
-            var result = _userService.Update(new UpdateDTO()
+            var result = await _userService.Update(new UpdateDTO()
             {
                 Email = email,
                 Fullname = fullname,
@@ -60,9 +83,14 @@ namespace EndPoint.Controllers
             });
             /*if (result.Status == Status.Success)
             {
-                
             }*/
-            return Ok(result);
+            if (result.Status == Status.Success)
+            {
+                return Ok(result);
+            }
+
+            return BadRequest(result);
+
         }
